# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para productos deportivos.",
        "title": "mlsport API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/products": {
            "get": {
                "description": "Devuelve una lista completa de productos registrados en el sistema.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener todos los productos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Product"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Permite registrar un nuevo producto en la base de datos.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Crear nuevo producto",
                "parameters": [
                    {
                        "description": "Producto a registrar",
                        "name": "producto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products/categories": {
            "get": {
                "description": "Retorna una lista de categorías derivadas de los productos registrados.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Listar categorías únicas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products/categories/{category}": {
            "get": {
                "description": "Filtra los productos según la categoría proporcionada.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener productos por categoría",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre de la categoría",
                        "name": "category",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Product"
                            }
                        }
                    }
                }
            }
        },
        "/products/dashboard": {
            "get": {
                "description": "Consulta combinada que obtiene productos y métricas en paralelo.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Dashboard de productos y métricas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products/metrics": {
            "get": {
                "description": "Devuelve métricas agregadas como total de productos, promedio de precios y stock acumulado.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Métricas de productos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Retorna la información detallada de un producto específico.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Consultar un producto por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del producto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza todos los campos de un producto existente con los nuevos valores proporcionados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Reemplazar producto existente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del producto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados del producto",
                        "name": "producto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Borra un producto según su ID.",
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar producto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del producto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Permite actualizar solo algunos campos de un producto existente.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar parcialmente un producto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del producto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a modificar (por ejemplo: stock, price)",
                        "name": "fields",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Product": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "objectID": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                }
            }
        }
    }
}